requirements.txt
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.28
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
pyyaml>=6.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
joblib>=1.3.0
tqdm>=4.66.0
